+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3ConsumerIT.java
@@ -14,9 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.s3.localstack;
-
-import java.util.List;
+package org.apache.camel.component.aws2.s3.integration;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
@@ -24,14 +22,10 @@ import org.apache.camel.Processor;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.s3.AWS2S3Constants;
-import org.apache.camel.component.aws2.s3.AWS2S3Operations;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.s3.model.S3Object;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class S3StreamUploadTimeoutLocalstackTest extends Aws2S3BaseTest {
+public class S3ConsumerIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -41,25 +35,37 @@ public class S3StreamUploadTimeoutLocalstackTest extends Aws2S3BaseTest {
 
     @Test
     public void sendIn() throws Exception {
-        result.expectedMessageCount(23);
+        result.expectedMessageCount(3);
 
-        for (int i = 0; i < 23; i++) {
-            template.sendBody("direct:stream1", "Andrea\n");
-        }
-
-        Thread.sleep(11000);
-        assertMockEndpointsSatisfied();
-
-        Exchange ex = template.request("direct:listObjects", new Processor() {
+        template.send("direct:putObject", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.listObjects);
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test.txt");
+                exchange.getIn().setBody("Test");
             }
         });
 
-        List<S3Object> resp = ex.getMessage().getBody(List.class);
-        assertEquals(1, resp.size());
+        template.send("direct:putObject", new Processor() {
+
+            @Override
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test1.txt");
+                exchange.getIn().setBody("Test1");
+            }
+        });
+
+        template.send("direct:putObject", new Processor() {
+
+            @Override
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test2.txt");
+                exchange.getIn().setBody("Test2");
+            }
+        });
+
+        Thread.sleep(10000);
+        assertMockEndpointsSatisfied();
     }
 
     @Override
@@ -67,14 +73,13 @@ public class S3StreamUploadTimeoutLocalstackTest extends Aws2S3BaseTest {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint1
-                        = "aws2-s3://mycamel-1?autoCreateBucket=true&streamingUploadMode=true&keyName=fileTest.txt&batchMessageNumber=25&namingStrategy=random&streamingUploadTimeout=10000";
+                String awsEndpoint = "aws2-s3://mycamel?autoCreateBucket=true";
 
-                from("direct:stream1").to(awsEndpoint1).to("mock:result");
+                from("direct:putObject").startupOrder(1).to(awsEndpoint).to("mock:result");
 
-                String awsEndpoint = "aws2-s3://mycamel-1?autoCreateBucket=true";
+                from("aws2-s3://mycamel?moveAfterRead=true&destinationBucket=camel-kafka-connector&autoCreateBucket=true&destinationBucketPrefix=RAW(movedPrefix)&destinationBucketSuffix=RAW(movedSuffix)")
+                        .startupOrder(2).log("${body}");
 
-                from("direct:listObjects").to(awsEndpoint);
             }
         };
     }
