+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3ConsumerIncludeBodyIT.java
@@ -14,22 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.s3.localstack;
+package org.apache.camel.component.aws2.s3.integration;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
-import org.apache.camel.ExchangePattern;
 import org.apache.camel.Processor;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.s3.AWS2S3Constants;
-import org.apache.camel.component.aws2.s3.AWS2S3Operations;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
+import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
-public class S3CreateDownloadLinkOperationLocalstackTest extends Aws2S3BaseTest {
+public class S3ConsumerIncludeBodyIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -37,37 +36,40 @@ public class S3CreateDownloadLinkOperationLocalstackTest extends Aws2S3BaseTest 
     @EndpointInject("mock:result")
     private MockEndpoint result;
 
-    @SuppressWarnings("unchecked")
     @Test
     public void sendIn() throws Exception {
-        result.expectedMessageCount(1);
+        result.expectedMessageCount(3);
 
-        template.send("direct:listBucket", new Processor() {
+        template.send("direct:putObject", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.listBuckets);
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test.txt");
+                exchange.getIn().setBody("Test");
             }
         });
 
-        template.send("direct:addObject", ExchangePattern.InOnly, new Processor() {
+        template.send("direct:putObject", new Processor() {
+
+            @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "CamelUnitTest2");
-                exchange.getIn().setBody("This is my bucket content.");
-                exchange.getIn().removeHeader(AWS2S3Constants.S3_OPERATION);
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test1.txt");
+                exchange.getIn().setBody("Test1");
             }
         });
 
-        Exchange ex1 = template.request("direct:createDownloadLink", new Processor() {
+        template.send("direct:putObject", new Processor() {
+
+            @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "CamelUnitTest2");
-                exchange.getIn().setHeader(AWS2S3Constants.BUCKET_NAME, "mycamel2");
-                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.createDownloadLink);
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test2.txt");
+                exchange.getIn().setBody("Test2");
             }
         });
 
-        assertNotNull(ex1.getMessage().getBody());
         assertMockEndpointsSatisfied();
+        assertEquals(3, result.getExchanges().size());
+        assertNotNull(result.getExchanges().get(0).getMessage().getBody());
     }
 
     @Override
@@ -75,14 +77,12 @@ public class S3CreateDownloadLinkOperationLocalstackTest extends Aws2S3BaseTest 
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint = "aws2-s3://mycamel2?autoCreateBucket=true";
+                String awsEndpoint = "aws2-s3://mycamel?autoCreateBucket=true";
 
-                from("direct:listBucket").to(awsEndpoint);
+                from("direct:putObject").startupOrder(1).to(awsEndpoint).to("mock:result");
 
-                from("direct:addObject").to(awsEndpoint);
-
-                from("direct:createDownloadLink").to(awsEndpoint + "&accessKey=xxx&secretKey=yyy&region=eu-west-1")
-                        .to("mock:result");
+                from("aws2-s3://mycamel?moveAfterRead=true&destinationBucket=camel-kafka-connector&autoCreateBucket=true&destinationBucketPrefix=RAW(movedPrefix)&destinationBucketSuffix=RAW(movedSuffix)&includeBody=false")
+                        .startupOrder(2).to("mock:result");
 
             }
         };
