+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3CreateDownloadLinkOperationIT.java
@@ -14,20 +14,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.s3.localstack;
-
-import java.io.File;
+package org.apache.camel.component.aws2.s3.integration;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
+import org.apache.camel.ExchangePattern;
 import org.apache.camel.Processor;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.s3.AWS2S3Constants;
+import org.apache.camel.component.aws2.s3.AWS2S3Operations;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
 
-public class S3MultipartUploadOperationLocalstackTest extends Aws2S3BaseTest {
+import static org.junit.Assert.assertNotNull;
+
+public class S3CreateDownloadLinkOperationIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -35,19 +37,36 @@ public class S3MultipartUploadOperationLocalstackTest extends Aws2S3BaseTest {
     @EndpointInject("mock:result")
     private MockEndpoint result;
 
+    @SuppressWarnings("unchecked")
     @Test
     public void sendIn() throws Exception {
         result.expectedMessageCount(1);
 
-        template.send("direct:putObject", new Processor() {
+        template.send("direct:listBucket", new Processor() {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "empty.txt");
-                exchange.getIn().setBody(new File("src/test/resources/empty.txt"));
+                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.listBuckets);
             }
         });
 
+        template.send("direct:addObject", ExchangePattern.InOnly, new Processor() {
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "CamelUnitTest2");
+                exchange.getIn().setBody("This is my bucket content.");
+                exchange.getIn().removeHeader(AWS2S3Constants.S3_OPERATION);
+            }
+        });
+
+        Exchange ex1 = template.request("direct:createDownloadLink", new Processor() {
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "CamelUnitTest2");
+                exchange.getIn().setHeader(AWS2S3Constants.BUCKET_NAME, "mycamel2");
+                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.createDownloadLink);
+            }
+        });
+
+        assertNotNull(ex1.getMessage().getBody());
         assertMockEndpointsSatisfied();
     }
 
@@ -56,9 +75,14 @@ public class S3MultipartUploadOperationLocalstackTest extends Aws2S3BaseTest {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint = "aws2-s3://mycamel?multiPartUpload=true&autoCreateBucket=true";
+                String awsEndpoint = "aws2-s3://mycamel2?autoCreateBucket=true";
 
-                from("direct:putObject").to(awsEndpoint).to("mock:result");
+                from("direct:listBucket").to(awsEndpoint);
+
+                from("direct:addObject").to(awsEndpoint);
+
+                from("direct:createDownloadLink").to(awsEndpoint + "&accessKey=xxx&secretKey=yyy&region=eu-west-1")
+                        .to("mock:result");
 
             }
         };
