+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3ListObjectsOperationIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.s3.localstack;
+package org.apache.camel.component.aws2.s3.integration;
 
 import java.util.List;
 
@@ -32,7 +32,7 @@ import software.amazon.awssdk.services.s3.model.S3Object;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class S3ListObjectsOperationLocalstackTest extends Aws2S3BaseTest {
+public class S3ListObjectsOperationIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
