+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3MultipartUploadOperationIT.java
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.aws2.s3.localstack;
+package org.apache.camel.component.aws2.s3.integration;
 
-import java.util.List;
+import java.io.File;
 
 import org.apache.camel.EndpointInject;
 import org.apache.camel.Exchange;
@@ -24,14 +24,10 @@ import org.apache.camel.Processor;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.aws2.s3.AWS2S3Constants;
-import org.apache.camel.component.aws2.s3.AWS2S3Operations;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.services.s3.model.S3Object;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-public class S3SimpleUploadOperationLocalstackTest extends Aws2S3BaseTest {
+public class S3MultipartUploadOperationIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -47,23 +43,10 @@ public class S3SimpleUploadOperationLocalstackTest extends Aws2S3BaseTest {
 
             @Override
             public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "camel.txt");
-                exchange.getIn().setBody("Camel rocks!");
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "empty.txt");
+                exchange.getIn().setBody(new File("src/test/resources/empty.txt"));
             }
         });
-
-        template.request("direct:listObjects", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.BUCKET_NAME, "mycamel");
-                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.listObjects);
-            }
-        });
-
-        List<S3Object> resp = result.getExchanges().get(0).getMessage().getBody(List.class);
-        assertEquals(1, resp.size());
-        assertEquals("camel.txt", resp.get(0).key());
 
         assertMockEndpointsSatisfied();
     }
@@ -73,11 +56,9 @@ public class S3SimpleUploadOperationLocalstackTest extends Aws2S3BaseTest {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint = "aws2-s3://mycamel?autoCreateBucket=true";
+                String awsEndpoint = "aws2-s3://mycamel?multiPartUpload=true&autoCreateBucket=true";
 
-                from("direct:putObject").to(awsEndpoint);
-
-                from("direct:listObjects").to(awsEndpoint).to("mock:result");
+                from("direct:putObject").to(awsEndpoint).to("mock:result");
 
             }
         };
