+++ components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/integration/S3SimpleUploadOperationIT.java
@@ -16,28 +16,22 @@
  */
 package org.apache.camel.component.aws2.s3.integration;
 
-import org.apache.camel.BindToRegistry;
+import java.util.List;
+
 import org.apache.camel.EndpointInject;
+import org.apache.camel.Exchange;
+import org.apache.camel.Processor;
 import org.apache.camel.ProducerTemplate;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.aws2.s3.AWS2S3Constants;
+import org.apache.camel.component.aws2.s3.AWS2S3Operations;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.apache.camel.test.junit5.CamelTestSupport;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
-import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
-import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.s3.S3Client;
+import software.amazon.awssdk.services.s3.model.S3Object;
 
-@Disabled("Must be manually tested. Provide your own accessKey and secretKey!")
-public class S3StreamUploadOperationIntegrationTest extends CamelTestSupport {
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-    @BindToRegistry("amazonS3Client")
-    S3Client client
-            = S3Client.builder()
-                    .credentialsProvider(StaticCredentialsProvider.create(
-                            AwsBasicCredentials.create("xxxx", "yyyy")))
-                    .region(Region.EU_WEST_1).build();
+public class S3SimpleUploadOperationIT extends Aws2S3Base {
 
     @EndpointInject
     private ProducerTemplate template;
@@ -47,13 +41,30 @@ public class S3StreamUploadOperationIntegrationTest extends CamelTestSupport {
 
     @Test
     public void sendIn() throws Exception {
-        result.expectedMessageCount(1000);
+        result.expectedMessageCount(1);
 
-        for (int i = 0; i < 1000; i++) {
-            template.sendBody("direct:stream1", "Andrea\n");
-        }
+        template.send("direct:putObject", new Processor() {
 
-        Thread.sleep(30000);
+            @Override
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.KEY, "camel.txt");
+                exchange.getIn().setBody("Camel rocks!");
+            }
+        });
+
+        template.request("direct:listObjects", new Processor() {
+
+            @Override
+            public void process(Exchange exchange) throws Exception {
+                exchange.getIn().setHeader(AWS2S3Constants.BUCKET_NAME, "mycamel");
+                exchange.getIn().setHeader(AWS2S3Constants.S3_OPERATION, AWS2S3Operations.listObjects);
+            }
+        });
+
+        List<S3Object> resp = result.getExchanges().get(0).getMessage().getBody(List.class);
+        assertEquals(1, resp.size());
+        assertEquals("camel.txt", resp.get(0).key());
+
         assertMockEndpointsSatisfied();
     }
 
@@ -62,12 +73,13 @@ public class S3StreamUploadOperationIntegrationTest extends CamelTestSupport {
         return new RouteBuilder() {
             @Override
             public void configure() throws Exception {
-                String awsEndpoint1
-                        = "aws2-s3://mycamel-1?autoCreateBucket=true&streamMode=true&keyName=fileTest.txt&batchMessageNumber=25&namingStrategy=random";
+                String awsEndpoint = "aws2-s3://mycamel?autoCreateBucket=true";
 
-                from("direct:stream1").toD(awsEndpoint1).to("mock:result");
+                from("direct:putObject").to(awsEndpoint);
+
+                from("direct:listObjects").to(awsEndpoint).to("mock:result");
+
             }
         };
     }
-
 }
