--- components/camel-aws/camel-aws2-s3/src/test/java/org/apache/camel/component/aws2/s3/localstack/S3ConsumerLocalstackIncludeBodyTest.java
@@ -1,90 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.camel.component.aws2.s3.localstack;
-
-import org.apache.camel.EndpointInject;
-import org.apache.camel.Exchange;
-import org.apache.camel.Processor;
-import org.apache.camel.ProducerTemplate;
-import org.apache.camel.builder.RouteBuilder;
-import org.apache.camel.component.aws2.s3.AWS2S3Constants;
-import org.apache.camel.component.mock.MockEndpoint;
-import org.junit.jupiter.api.Test;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-
-public class S3ConsumerLocalstackIncludeBodyTest extends Aws2S3BaseTest {
-
-    @EndpointInject
-    private ProducerTemplate template;
-
-    @EndpointInject("mock:result")
-    private MockEndpoint result;
-
-    @Test
-    public void sendIn() throws Exception {
-        result.expectedMessageCount(3);
-
-        template.send("direct:putObject", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test.txt");
-                exchange.getIn().setBody("Test");
-            }
-        });
-
-        template.send("direct:putObject", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test1.txt");
-                exchange.getIn().setBody("Test1");
-            }
-        });
-
-        template.send("direct:putObject", new Processor() {
-
-            @Override
-            public void process(Exchange exchange) throws Exception {
-                exchange.getIn().setHeader(AWS2S3Constants.KEY, "test2.txt");
-                exchange.getIn().setBody("Test2");
-            }
-        });
-
-        assertMockEndpointsSatisfied();
-        assertEquals(3, result.getExchanges().size());
-        assertNotNull(result.getExchanges().get(0).getMessage().getBody());
-    }
-
-    @Override
-    protected RouteBuilder createRouteBuilder() throws Exception {
-        return new RouteBuilder() {
-            @Override
-            public void configure() throws Exception {
-                String awsEndpoint = "aws2-s3://mycamel?autoCreateBucket=true";
-
-                from("direct:putObject").startupOrder(1).to(awsEndpoint).to("mock:result");
-
-                from("aws2-s3://mycamel?moveAfterRead=true&destinationBucket=camel-kafka-connector&autoCreateBucket=true&destinationBucketPrefix=RAW(movedPrefix)&destinationBucketSuffix=RAW(movedSuffix)&includeBody=false")
-                        .startupOrder(2).to("mock:result");
-
-            }
-        };
-    }
-}
